add Weapon & Gear Page (#2)
